requirements.txt
accelerate
bitsandbytes
datasets
deepspeed==0.14.4
einops
flash-attn==2.5.8
jsonlines
loralib
optimum
packaging
peft
dm-launchpad[tensorflow]
torchmetrics
tqdm
llm-blender==0.0.2
transformers==4.43.3
vllm==0.4.2
transformers_stream_generator
wandb
wheel
ml_collections
pyarrow>=0.6.1,<12
python-Levenshtein==0.25.1
mosec
